requirement.txt
absl-py==1.2.0	
addict==2.4.0	
anyconfig==0.9.10	
backcall==0.2.0	
ca-certificates==2022.6.15	
cachetools==4.2.4	
certifi==2022.6.15	
charset-normalizer==2.0.12	
click==7.1.2	
colorama==0.4.5	
cycler==0.11.0	
dataclasses==0.6	
decorator==4.4.2	
fire==0.4.0	
fonttools==4.37.0	
future==0.18.2	
google-auth==1.35.0	
google-auth-oauthlib==0.4.6	
grpcio==1.47.0	
idna==3.3	
imageio==2.21.1	
imgaug==0.4.0	
importlib-metadata==4.8.3	
ipython==7.34.0	
ipython-genutils==0.2.0	
jedi==0.18.1	
kiwisolver	==1.4.4	
libsqlite==3.39.2	
lmdb==0.97	
markdown==3.4.1	
markupsafe==2.1.1	
matplotlib==3.5.3	
matplotlib-inline==0.1.6	
natsort==8.1.0	
networkx==2.6.3	
numpy==1.21.6	
oauthlib==3.2.0	
opencv-python==4.6.0.66	
openssl==3.0.5	
packaging==21.3	
params==0.9.0	
parso==0.8.3	
pexpect==4.8.0	
pickleshare==0.7.5	
pillow==9.2.0	
pip==22.2.2	
polygon3==3.0.9.1	
prompt-toolkit==3.0.30	
protobuf==3.19.4	
ptyprocess==0.7.0	
pyasn1==0.4.8	
pyasn1-modules==0.2.8	
pyclipper==1.3.0.post3	
pygments	==2.11.2	
pyparsing	==3.0.9	
pyqt5==5.15.4	
pyqt5-plugins==5.15.4.2.2	
pyqt5-qt5==5.15.2	
pyqt5-sip==12.11.0	
pyqt5-tools==5.15.4.3.2	
python==3.7.12	
python-dateutil==2.8.2	
python-dotenv==0.20.0	
python_abi==3.7	
pywavelets==1.3.0	
pyyaml==5.2	
qt5-applications==5.15.2.2.2	
qt5-tools==5.15.2.1.2	
requests==2.28.1	
requests-oauthlib==1.3.1	
rsa==4.9	
scikit-image==0.19.3	
scipy==1.7.3	
setuptools==65.2.0	
shapely==1.8.4	
six==1.16.0	
sqlite==3.39.2	
tensorboard==2.10.0	
tensorboard-data-server==0.6.1	
tensorboard-plugin-wit==1.8.1	
termcolor==1.1.0	
tifffile==2021.11.2	
torch==1.7.0+cu110	
torchaudio==0.7.0	
torchvision==0.8.1+cu110	
tqdm==4.40.1	
traitlets==5.3.0	
typing-extensions==4.3.0	
ucrt==10.0.20348.0	
urllib3==1.26.12	
vc==14.2	
vs2015_runtime==14.29.30037	
wcwidth==0.2.5	
werkzeug	==2.2.2	
wheel==0.37.1	
zipp==3.8.1	